anomaly_engine/requirements.txt
asyncpg==0.29.0
redis==5.0.4
python-dotenv==1.0.1
scikit-learn==1.4.2
numpy==1.26.4
httpx==0.27.0
